requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
dnspython==2.4.2
colorama==0.4.6
tqdm==4.66.1
pyyaml==6.0.1
selenium==4.15.2
fake-useragent==1.4.0
urllib3==2.0.7
certifi==2023.7.22
python-whois==0.8.0
argparse==1.4.0
validators==0.22.0
